requirements.txt
beautifulsoup4==4.9.0
boto3==1.7.74
tomd==0.1.3
pynacl==1.4.0
requests==2.25.1
urllib3==1.26.3